refactor: move Years component to Years.tsx file
